+++ example/lib/views/tabs/tabs.dart
@@ -1,0 +1,9 @@
+library;
+
+import 'package:flutter/material.dart';
+import 'package:layrz_icons/layrz_icons.dart';
+import 'package:layrz_theme/layrz_theme.dart';
+import 'package:layrz_theme_example/store/store.dart';
+
+part 'src/basic.dart';
+part 'src/advanced.dart';